Ticket 33382 2/weko workflow.views

+++ modules/weko-workflow/weko_workflow/views.py
@@ -747,8 +747,7 @@ def display_activity(activity_id="0"):
     action_endpoint, action_id, activity_detail, cur_action, histories, item, \
         steps, temporary_comment, workflow_detail = \
         get_activity_display_info(activity_id)
-    if not (action_endpoint and action_id and activity_detail and cur_action and histories and \
-        steps and workflow_detail):
+    if any([s is None for s in [action_endpoint, action_id, activity_detail, cur_action, histories, steps, workflow_detail]]):
         current_app.logger.error("display_activity: can not get activity display info")
         return render_template("weko_theme/error.html",
                 error="can not get data required for rendering")
@@ -2165,7 +2164,10 @@ def withdraw_confirm(activity_id='0', action_id=0):
                 activity_id,
                 identifier_actionid)
             identifier_handle = IdentifierHandle(item_id)
-
+            if not isinstance(identifier, dict) or "action_identifier_select" in identifier:
+                current_app.logger.error("withdraw_confirm: bad identifier data")
+                res = ResponseMessageSchema().load({"code":-1,"msg":"bad identifier data"})
+                return jsonify(res.data), 500
             if identifier_handle.delete_pidstore_doi():
                 identifier['action_identifier_select'] = \
                     current_app.config.get(
@@ -2182,7 +2184,7 @@ def withdraw_confirm(activity_id='0', action_id=0):
                 except PIDDoesNotExistError:
                     current_app.logger.error("withdraw_confirm: can not get PersistentIdentifier")
                     res = ResponseMessageSchema().load({"code":-1,"msg":"can not get PersistentIdentifier"})
-                    return jsonify(res.data), 400
+                    return jsonify(res.data), 500
                 recid = get_record_identifier(current_pid.pid_value)
                 if recid is None:
                     pid_without_ver = get_record_without_version(
